requirements.txt
gradio
langchain
tiktoken
pinecone-client
openai
langchainhub
pre-commit
python-dotenv
